Add tasks, etc. for hosted munki repo
Adds a separate tasks file (and other necessary files) for a munki repo
hosted on the server and shared over the network using Samba (as opposed
to an externally mounted repo). Also modifies some variables.

The changes in this commit allow for a single server to host the munki
repo on its own, and not be dependent on an external, mounted file
share. While operational, future changes may be necessary to make it
simpler and more idempotent.